+++ anyblok_pyramid/bloks/authorization/authorization.py
@@ -11,8 +11,11 @@ from anyblok.column import Integer, String, Json
 from anyblok.relationship import Many2One
 from anyblok.field import JsonRelated
 from .exceptions import AuthorizationValidationException
-from pyramid.authorization import Allow, Deny, ALL_PERMISSIONS
 from sqlalchemy import or_
+try:
+    from pyramid.authorization import Allow, Deny, ALL_PERMISSIONS
+except ImportError:
+    from pyramid.security import Allow, Deny, ALL_PERMISSIONS
 
 
 Pyramid = Declarations.Model.Pyramid
@@ -202,7 +205,8 @@ class Authorization:
         :param kwargs: authorization fields
         """
         if not kwargs:
-            kwargs = {}
+            kwargs = {}  # pragma: no cover
+
         # pv: at some point adding index on this criteria may boost things
         # while setting authorizations
         authz = (