+++ anyblok_pyramid/security.py
@@ -6,7 +6,10 @@
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
 from pyramid.httpexceptions import HTTPUnauthorized
-from pyramid.authorization import Deny, Everyone, ALL_PERMISSIONS
+try:
+    from pyramid.authorization import Deny, Everyone, ALL_PERMISSIONS
+except ImportError:
+    from pyramid.security import Deny, Everyone, ALL_PERMISSIONS
 
 
 def group_finder(userid, request):
@@ -49,7 +52,8 @@ def AnyBlokResourceFactory(resource):
     """
     def __acl__(self):
         if not hasattr(self, 'registry'):
-            raise HTTPUnauthorized("ACL have not get AnyBlok registry")
+            raise HTTPUnauthorized(  # pragma: no cover
+                "ACL have not get AnyBlok registry")
 
         userid = self.request.authenticated_userid
         if userid: